+++ packages/parte-icons/src/Icons/WriteIcon.tsx
@@ -1,0 +1,29 @@
+import { convertIcon, IconProps } from "../common/Icon";
+function Icon({ size, ...rest }: Pick<IconProps, "style" | "color" | "size">) {
+  return (
+    <svg
+      width={size}
+      height={size}
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...rest}
+    >
+      <path
+        d="M1.77548 14.4999L4.60382 14.4999L14.9747 4.12904L12.1463 1.30063L1.77539 11.6715L1.77548 14.4999Z"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M9.31738 4.12906L12.1458 6.95749"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
+const WriteIcon = convertIcon(Icon, "write");
+export default WriteIcon;