+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @parte-ds/web
 
+## 0.1.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @parte-ds/icons@1.8.3
+
 ## 0.1.41
 
 ### Patch Changes